move irods_validation() to utils
