+++ solosis/commands/irods/iget_fastqs.py
@@ -6,7 +6,7 @@ import time
 import click
 import pandas as pd
 
-from solosis.utils import log_command
+from solosis.utils import irods_validation, log_command
 
 FASTQ_EXTENSIONS = [".fastq", ".fastq.gz"]
 
@@ -19,41 +19,6 @@ def spinner():
     while True:
         for frame in spinner_frames:
             yield frame
-
-
-def irods_validation():
-    """Run a command and handle specific output conditions."""
-    command = [
-        "iget",
-        "/seq/illumina/runs/48/48297/cellranger/cellranger720_count_48297_58_rBCN14591738_GRCh38-2020-A/web_summary.html",
-    ]
-
-    try:
-        # Run the command and capture stdout and stderr
-        result = subprocess.run(
-            command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-        )
-
-        # Check for the specific error in stderr
-        if (
-            "CAT_INVALID_AUTHENTICATION" in result.stderr
-            or "-827000 CAT_INVALID_USER" in result.stderr
-        ):
-            echo_message(
-                "run `iinit` before re-running this solosis command.",
-                "error",
-            )
-            sys.exit(1)  # Exit with error status 1
-
-        # If no error, command executed successfully
-        echo_message("Command executed successfully.", "success")
-
-    except FileNotFoundError:
-        echo_message(
-            "iRODS not loaded. please run `module load cellgen/irods` before re-running this solosis command.",
-            "error",
-        )
-        sys.exit(1)
 
 
 @click.command("iget-fastqs")
